scratch that, try this instead
not tracking the HDF5 file didn't work because it seems files in the
repo don't persist between processes on Heroku. the plan was to create
the file on the server with pivot_tables.py, so that it would be created
using appropriate versions of everything and then could be read by
workers. that's not going to work now because the file won't stick
around. instead, let's just try to use the same version of python that
I'm using at home.